requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1
httpx==0.26.0
beautifulsoup4==4.12.3
jinja2==3.1.3